Refactored folder structure, updated go.mod
