--- go.mod
@@ -1,15 +1,0 @@
-module github.com/c3re/can2mqtt
-
-go 1.18
-
-require (
-	github.com/brutella/can v0.0.2
-	github.com/eclipse/paho.mqtt.golang v1.4.1
-)
-
-require (
-	github.com/gorilla/websocket v1.4.2 // indirect
-	golang.org/x/net v0.17.0 // indirect
-	golang.org/x/sync v0.0.0-20210220032951-036812b2e83c // indirect
-	golang.org/x/sys v0.13.0 // indirect
-)
