+++ src/go.mod
@@ -1,0 +1,15 @@
+module github.com/c3re/can2mqtt
+
+go 1.22.2
+
+require (
+	github.com/brutella/can v0.0.2
+	github.com/eclipse/paho.mqtt.golang v1.4.3
+)
+
+require (
+	github.com/gorilla/websocket v1.5.0 // indirect
+	golang.org/x/net v0.8.0 // indirect
+	golang.org/x/sync v0.1.0 // indirect
+	golang.org/x/sys v0.6.0 // indirect
+)
